Move ClientState out of platform; reusable diagnostic event
