+++ src/main/java/org/orecruncher/dsurround/eventing/handlers/BlockUpdateHandler.java
@@ -4,7 +4,7 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.block.state.BlockState;
 import org.orecruncher.dsurround.eventing.ClientEventHooks;
-import org.orecruncher.dsurround.lib.platform.events.ClientState;
+import org.orecruncher.dsurround.eventing.ClientState;
 
 import java.util.Collection;
 import java.util.HashSet;
@@ -44,7 +44,9 @@ public class BlockUpdateHandler {
             return Optional.empty();
 
         // Need to expand out the updates to adjacent blocks.  A state change
-        // of a block may affect how the adjacent blocks are handled.
+        // of a block may affect how the adjacent blocks are handled. Can't rely on
+        // neighbor state changes since the effects the mod produces are virtual
+        // and do not exist server side.
         Set<BlockPos> updates = new HashSet<>();
         for (final BlockPos center : updatedPositions) {
             for (int i = -1; i < 2; i++)
