+++ src/main/java/org/orecruncher/dsurround/fabric/services/ClientEventRegistrationsImpl.java
@@ -2,7 +2,7 @@ package org.orecruncher.dsurround.fabric.services;
 
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayConnectionEvents;
 import org.orecruncher.dsurround.lib.platform.IClientEventRegistrations;
-import org.orecruncher.dsurround.lib.platform.events.ClientState;
+import org.orecruncher.dsurround.eventing.ClientState;
 
 public class ClientEventRegistrationsImpl implements IClientEventRegistrations {
