+++ src/main/java/org/orecruncher/dsurround/gui/overlay/plugins/ClientProfilerPlugin.java
@@ -8,7 +8,7 @@ import org.orecruncher.dsurround.gui.overlay.IDiagnosticPlugin;
 import org.orecruncher.dsurround.lib.events.HandlerPriority;
 import org.orecruncher.dsurround.lib.math.ITimer;
 import org.orecruncher.dsurround.lib.math.TimerEMA;
-import org.orecruncher.dsurround.lib.platform.events.ClientState;
+import org.orecruncher.dsurround.eventing.ClientState;
 
 public class ClientProfilerPlugin implements IDiagnosticPlugin {
