+++ src/main/java/org/orecruncher/dsurround/mixins/events/MixinMinecraftClient.java
@@ -1,7 +1,7 @@
 package org.orecruncher.dsurround.mixins.events;
 
 import net.minecraft.client.Minecraft;
-import org.orecruncher.dsurround.lib.platform.events.ClientState;
+import org.orecruncher.dsurround.eventing.ClientState;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
